Consolidate entity counts hooks
